Hide faster fetch op behaviour behind a flag
The new, faster `getOps` behaviour should be fine, but there may be
unforeseen corner cases where the underlying assumptions break down (eg
if consumers have been meddling with ops or snapshots outside of
ShareDB).

In order to remain conservative, and maintain data consistency in all
mainline cases, this change hides the faster behaviour behind a flag,
which defaults to off.

In order to make sure `sharedb-mongo` still acts correctly with this
flag enabled, we re-run all the tests again with the flag enabled.
